These should have been doubles.
git-svn-id: http://svn.code.sf.net/p/nitro-nitf/code/trunk@741 7dc3ea37-dbd3-4c13-bb9a-7ccd57e75167

+++ c/nitf/include/nitf/DateTime.h
@@ -126,7 +126,7 @@ NITFAPI(NITF_BOOL) nitf_DateTime_setSecond(nitf_DateTime *dateTime,
  * the remaining fields to match the new timeInMillis.
  */
 NITFAPI(NITF_BOOL) nitf_DateTime_setTimeInMillis(nitf_DateTime *dateTime,
-                                                 int timeInMillis,
+                                                 double timeInMillis,
                                                  nitf_Error *error);
 
 /*!
